Fix TypeScript type error in star-background

+++ components/main/star-background.tsx
@@ -13,7 +13,7 @@ import type { Points as PointsType } from "three";
 export const StarBackground = (props: PointsInstancesProps) => {
   const ref = useRef<PointsType | null>(null);
   const [sphere] = useState(() =>
-    random.inSphere(new Float32Array(5000), { radius: 1.2 }),
+    random.inSphere(new Float32Array(5000), { radius: 1.2 }) as Float32Array,
   );
 
   useFrame((_state, delta) => {
